+++ google/cloud/certificatemanager/quickstart/quickstart.cc
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/certificatemanager/ EDIT HERE .h"
-#include "google/cloud/project.h"
+#include "google/cloud/certificatemanager/certificate_manager_client.h"
 #include <iostream>
 #include <stdexcept>
 
@@ -24,16 +23,19 @@ int main(int argc, char* argv[]) try {
   }
 
   namespace certificatemanager = ::google::cloud::certificatemanager;
-  auto client = certificatemanager::Client(
-      certificatemanager::MakeConnection(/* EDIT HERE */));
+  auto client = certificatemanager::CertificateManagerClient(
+      certificatemanager::MakeCertificateManagerConnection());
 
-  auto const project = google::cloud::Project(argv[1]);
-  for (auto r : client.List /*EDIT HERE*/ (project.FullName())) {
-    if (!r) throw std::runtime_error(r.status().message());
+  auto const parent = std::string{"projects/"} + argv[1] + "/locations/-";
+  for (auto r : client.ListCertificates(parent)) {
+    if (!r) throw std::move(r).status();
     std::cout << r->DebugString() << "\n";
   }
 
   return 0;
+} catch (google::cloud::Status const& status) {
+  std::cerr << "google::cloud::Status thrown: " << status << "\n";
+  return 1;
 } catch (std::exception const& ex) {
   std::cerr << "Standard exception raised: " << ex.what() << "\n";
   return 1;