+++ CHANGELOG.md
@@ -123,8 +123,9 @@ may elicit a deprecation warning. See [#9086] for more details.
 We are happy to announce the following GA libraries.  Unless specifically noted,
 the APIs in these libraries are stable, and are ready for production use.
 
-- [Datastream API](/google/cloud/datastream/README.md)
+- [Certificate Manager API](/google/cloud/certificatemanager/README.md)
 - [Cloud Deploy API](/google/cloud/deploy/README.md)
+- [Datastream API](/google/cloud/datastream/README.md)
 
 In addition, these existing libraries are now GA: